Refactor api 폴더 위치 변경, 모듈화
